Moved love UI hooks to root UI and fixed modals not working
