+++ main.lua
@@ -22,8 +22,10 @@ function love.load(arg)
 	
 	--build ui
 	ui = require("ui.chaoshead"):new()
-	ui = require("ui.base.root"):new(ui)
-	ui:resize(love.graphics.getWidth(), love.graphics.getHeight())
+	uiRoot = require("ui.base.root"):new(ui)
+	uiRoot:hookIntoLove()
+	uiRoot:resize(love.graphics.getWidth(), love.graphics.getHeight())
+	
 	--bind ui and input
 	input.parseActions(settings.bindings)
 	input.inputActivated = function(...)
@@ -34,30 +36,9 @@ function love.load(arg)
 	end
 end
 
-function love.resize(w, h)
-	ui:resize(w,h)
-end
-
-function love.update(dt)
-	ui:update(dt)
-end
-function love.draw()
-	ui:draw()
-end
-
-function love.focus(focus)
-	ui:focus(focus)
-end
-function love.visible(visible)
-	ui:visible(visible)
-end
-
 function love.keypressed(key, scancode, isrepeat)
 	input.keypressed(key, scancode, isrepeat)
 	--ui:keypressed(key, scancode, isrepeat)
-end
-function love.textinput(text)
-	ui:textInput(text)
 end
 function love.keyreleased(key, scancode)
 	input.keyreleased(key, scancode)
@@ -70,10 +51,4 @@ end
 function love.mousereleased(x, y, button, isTouch, presses)
 	input.mousereleased(x,y, button, isTouch, presses)
 	--ui:mousereleased(x,y, button, isTouch)
-end
-function love.mousemoved(x, y, dx, dy)
-	ui:mouseMoved(x, y, dx, dy)
-end
-function love.wheelmoved(dx, dy)
-	ui:wheelMoved(dx,dy)
 end
